third_party/blink/web_tests/storage/indexeddb/intversion-abort-in-initial-upgradeneeded-expected.txt

Test that an abort in the initial upgradeneeded sets version back to 0

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "intversion-abort-in-initial-upgradeneeded.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname, 2)
request.onupgradeneeded = upgradeNeeded
request.onerror = onError


upgradeNeeded():
PASS db.version is 2
transaction.abort()


onAbort():
PASS event.target.db.version is 0
PASS request.transaction is non-null.


onError():
PASS request is event.target
PASS event.target.result is undefined.
PASS event.target.error is non-null.
PASS event.target.error.name is "AbortError"
PASS request.transaction is null
PASS successfullyParsed is true

TEST COMPLETE
